v2

+++ programs/vyper-calculator/src/lib.rs
@@ -8,27 +8,27 @@ pub mod vyper_calculator {
 
     pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
         ctx.accounts.state.authority = ctx.accounts.authority.key();
-        ctx.accounts.state.value = 0;
+        ctx.accounts.state.value = 0f64;
         Ok(())
     }
 
     pub fn sum(ctx: Context<OpContext>, a: u64, b: u64) -> Result<()> {
-        ctx.accounts.state.value = a + b;
+        ctx.accounts.state.value = (a as f64) + (b as f64);
         Ok(())
     }
 
     pub fn sub(ctx: Context<OpContext>, a: u64, b: u64) -> Result<()> {
-        ctx.accounts.state.value = a - b;
+        ctx.accounts.state.value = (a as f64) - (b as f64);
         Ok(())
     }
 
     pub fn mul(ctx: Context<OpContext>, a: u64, b: u64) -> Result<()> {
-        ctx.accounts.state.value = a * b;
+        ctx.accounts.state.value =(a as f64) * (b as f64);
         Ok(())
     }
 
     pub fn div(ctx: Context<OpContext>, a: u64, b: u64) -> Result<()> {
-        ctx.accounts.state.value = a / b;
+        ctx.accounts.state.value = (a as f64) / (b as f64);
         Ok(()) 
     }
 
@@ -37,13 +37,13 @@ pub mod vyper_calculator {
 #[account]
 pub struct State {
     pub authority: Pubkey,
-    pub value: u64
+    pub value: f64
 }
 
 impl State {
     pub const LEN: usize = 8 + // discriminator
     32 + // pub authority: Pubkey,
-    8;  // pub value: u64
+    8;  // pub value: f64
 }
 
 #[derive(Accounts)]
